Codsoft_taskno1.java
The Random Number Generator Java program is designed to generate random numbers based on specified criteria and requirements. It provides a user-friendly interface and versatile functionality for generating random numbers in a Java environment.

Feel free to explore the code, contribute to its development, or utilize it for your own projects. Your feedback and suggestions are always welcome as I continue to enhance this program as part of my internship experience at Codsoft.